chore(tests): use `usize` for unassigned nodes

 

See merge request dfinity-lab/public/ic!16371

+++ rs/tests/src/orchestrator/subnet_recovery_nns_failover.rs
@@ -66,7 +66,7 @@ pub fn setup(env: TestEnv) {
                 .with_dkg_interval_length(Height::from(DKG_INTERVAL))
                 .add_nodes(1),
         )
-        .with_unassigned_nodes(SUBNET_SIZE as i32)
+        .with_unassigned_nodes(SUBNET_SIZE)
         .setup_and_start(&env)
         .expect("failed to setup IC under test");
